# swaggerDesign.json
{
    "swagger": "2.0",
    "info": {
      "description": "This is an API server for a collaborative task management application.",
      "version": "1.0.0",
      "title": "Collaborative Task Management API",
      "contact": {
        "email": "your_email@example.com"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "task-cse341-final.onrender.com",
    "basePath": "/",
...
    "schemes": ["https"],
    "paths": {
...
      "/task": {
        "post": {
          "tags": ["task"],
          "summary": "Create task",
          "description": "This endpoint is used to create a new task.",
          "operationId": "createTask",
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Task object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Task"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Task created successfully",
              "schema": {
                "$ref": "#/definitions/Task"
              }
            },
            "400": {
              "description": "Invalid task data provided"
            }
          }
        }
      },
      "/team": {
        "post": {
          "tags": ["team"],
          "summary": "Create team",
          "description": "This endpoint is used to create a new team.",
          "operationId": "createTeam",
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Team object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Team"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Team created successfully",
              "schema": {
                "$ref": "#/definitions/Team"
              }
            },
            "400": {
              "description": "Invalid team data provided"
            }
          }
        }
      }
    },
    "definitions": {
...
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["todo", "in_progress", "done"]
          }
        }
      },
      "Team": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }